+++ backend/src/main/java/io/debezium/configserver/service/ConnectorIntegratorBase.java
@@ -1,0 +1,146 @@
+package io.debezium.configserver.service;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+import org.apache.kafka.common.config.Config;
+import org.apache.kafka.common.config.ConfigDef;
+import org.apache.kafka.common.config.ConfigDef.ConfigKey;
+import org.apache.kafka.common.config.ConfigValue;
+import org.apache.kafka.connect.source.SourceConnector;
+
+import io.debezium.config.Configuration;
+import io.debezium.config.Field;
+import io.debezium.configserver.model.ConnectionValidationResult;
+import io.debezium.configserver.model.ConnectorProperty;
+import io.debezium.configserver.model.ConnectorType;
+import io.debezium.configserver.model.GenericValidationResult;
+import io.debezium.configserver.model.PropertiesValidationResult;
+import io.debezium.configserver.model.PropertyValidationResult;
+import io.debezium.connector.postgresql.PostgresConnectorConfig;
+
+public abstract class ConnectorIntegratorBase implements ConnectorIntegrator {
+
+    protected abstract ConnectorDescriptor getConnectorDescriptor();
+
+    protected abstract SourceConnector getConnector();
+
+    @Override
+    public ConnectorType getDescriptor() {
+        ConnectorDescriptor descriptor = getConnectorDescriptor();
+        SourceConnector instance = getConnector();
+
+        List<ConnectorProperty> properties = instance.config()
+            .configKeys()
+            .values()
+            .stream()
+            .map(this::toConnectorProperty)
+            .collect(Collectors.toList());
+
+        return new ConnectorType(
+                descriptor.id,
+                instance.getClass().getName(),
+                descriptor.name,
+                instance.version(),
+                descriptor.enabled,
+                properties
+        );
+    }
+
+    @Override
+    public ConnectionValidationResult validateConnection(Map<String, String> properties) {
+        SourceConnector instance = getConnector();
+
+        try {
+            Config result = instance.validate(properties);
+            List<PropertyValidationResult> propertyResults = toPropertyValidationResults(result);
+
+            return propertyResults.isEmpty() ? ConnectionValidationResult.valid() : ConnectionValidationResult.invalid(propertyResults);
+        }
+        catch(Exception e) {
+            return ConnectionValidationResult.invalid(Collections.emptyList(), Collections.singletonList(new GenericValidationResult(e.getMessage(), traceAsString(e))));
+        }
+    }
+
+    @Override
+    public PropertiesValidationResult validateProperties(Map<String, String> properties) {
+        List<Field> fields = new ArrayList<>();
+        PostgresConnectorConfig.ALL_FIELDS.forEach(field -> {
+            if (properties.containsKey(field.name())) {
+                fields.add(field);
+            }
+        });
+
+        Configuration config = Configuration.from(properties);
+        Map<String, ConfigValue> results = config.validate(Field.setOf(fields));
+        Config result = new Config(new ArrayList<>(results.values()));
+
+        List<PropertyValidationResult> propertyResults = toPropertyValidationResults(result);
+
+        return propertyResults.isEmpty() ? PropertiesValidationResult.valid() : PropertiesValidationResult.invalid(propertyResults);
+    }
+
+    private List<PropertyValidationResult> toPropertyValidationResults(Config result) {
+        List<PropertyValidationResult> propertyResults = result.configValues()
+                .stream()
+                .filter(cv -> !cv.errorMessages().isEmpty())
+                .map(cv -> new PropertyValidationResult(cv.name(), cv.errorMessages().get(0)))
+                .collect(Collectors.toList());
+        return propertyResults;
+    }
+
+    private String traceAsString(Exception e) {
+        return e.getStackTrace() != null && e.getStackTrace().length > 0 ? Arrays.toString(e.getStackTrace()) : null;
+    }
+
+    private ConnectorProperty toConnectorProperty(ConfigKey configKey) {
+        return new ConnectorProperty(
+                configKey.name,
+                configKey.displayName,
+                configKey.documentation,
+                toConnectorPropertyType(configKey.type()),
+                configKey.defaultValue
+        );
+    }
+
+    private ConnectorProperty.Type toConnectorPropertyType(ConfigDef.Type type) {
+        switch(type) {
+            case BOOLEAN:
+                return ConnectorProperty.Type.BOOLEAN;
+            case CLASS:
+                return ConnectorProperty.Type.CLASS;
+            case DOUBLE:
+                return ConnectorProperty.Type.DOUBLE;
+            case INT:
+                return ConnectorProperty.Type.INT;
+            case LIST:
+                return ConnectorProperty.Type.LIST;
+            case LONG:
+                return ConnectorProperty.Type.LONG;
+            case PASSWORD:
+                return ConnectorProperty.Type.PASSWORD;
+            case SHORT:
+                return ConnectorProperty.Type.SHORT;
+            case STRING:
+                return ConnectorProperty.Type.STRING;
+            default:
+                throw new IllegalArgumentException("Unsupported property type: " + type);
+        }
+    }
+
+    public static class ConnectorDescriptor {
+        public String id;
+        public String name;
+        public boolean enabled;
+
+        public ConnectorDescriptor(String id, String name, boolean enabled) {
+            this.id = id;
+            this.name = name;
+            this.enabled = enabled;
+        }
+    }
+}
